Move TooltipContent to a separate component
